requirements.txt
dnspython
pycodestyle
pylint
pyOpenSSL
pyYAML
requests
